Issue#225: StringToContactTypeConverter moved to converter package
