+++ home-application/src/main/java/com/softserveinc/ita/homeproject/application/converter/StringToContactTypeConverter.java
@@ -1,4 +1,4 @@
-package com.softserveinc.ita.homeproject.application.provider;
+package com.softserveinc.ita.homeproject.application.converter;
 
 import javax.ws.rs.ext.Provider;
